+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * *[v2.2.0-rc1]*
   * [FEATURE]: Add support for tracking a video buffering
   * [FEATURE]: Add support for tracking a delay with every event type
+  * [REFACTOR]: Clean up the code used to communicate outside of the
+    player `<iframe>`.
 
 ## v2.1.3 (December 9, 2015)
 * *[v2.1.3-rc2]*